# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  clerkUserId: string;
  name: string;
  stream: string;
  college: string;
  state: string;
  category: string;
  financialBackground: string;
  careerInterest: string;
  yearOfStudy: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileInput {
  name: string;
  stream: string;
  college: string;
  state: string;
  category: string;
  financialBackground: string;
  careerInterest: string;
  yearOfStudy: string;
}

export interface Scholarship {
  id: number;
  name: string;
  provider: string;
  amount: string;
  deadline: string;
  applyUrl: string;
  eligibility: string;
  category: string;
  /** @nullable */
  state?: string | null;
  description: string;
}

export interface Exam {
  id: number;
  name: string;
  conductedBy: string;
  examDate: string;
  registrationDeadline: string;
  officialUrl: string;
  stream: string;
  description: string;
  careerRelevance: string;
}

export interface Resource {
  id: number;
  title: string;
  platform: string;
  url: string;
  type: string;
  stream: string;
  language: string;
  description: string;
}

export interface Mentor {
  id: number;
  name: string;
  title: string;
  college: string;
  careerDomain: string;
  background: string;
  contactUrl: string;
  availability: string;
  bio: string;
}

export interface SaveResult {
  ok: boolean;
}

export interface UpcomingDeadline {
  examId: number;
  examName: string;
  registrationDeadline: string;
  daysRemaining: number;
  officialUrl: string;
  stream: string;
}

export interface DashboardSummary {
  scholarshipsCount: number;
  examsCount: number;
  resourcesCount: number;
  mentorsCount: number;
  hasProfile: boolean;
}