# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  scholarshipsCount: number;
  examsCount: number;
  resourcesCount: number;
  mentorsCount: number;
  hasProfile: boolean;
}