# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: number;
  name: string;
  conductedBy: string;
  examDate: string;
  registrationDeadline: string;
  officialUrl: string;
  stream: string;
  description: string;
  careerRelevance: string;
}