# lib/api-zod/src/generated/types/mentor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Mentor {
  id: number;
  name: string;
  title: string;
  college: string;
  careerDomain: string;
  background: string;
  contactUrl: string;
  availability: string;
  bio: string;
}