# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  clerkUserId: string;
  name: string;
  stream: string;
  college: string;
  state: string;
  category: string;
  financialBackground: string;
  careerInterest: string;
  yearOfStudy: string;
  createdAt: string;
  updatedAt: string;
}