# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  title: string;
  platform: string;
  url: string;
  type: string;
  stream: string;
  language: string;
  description: string;
}