# lib/api-zod/src/generated/types/upcomingDeadline.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduPath API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpcomingDeadline {
  examId: number;
  examName: string;
  registrationDeadline: string;
  daysRemaining: number;
  officialUrl: string;
  stream: string;
}